usef-build/usef-core/usef-core-transport/src/test/java/energy/usef/core/repository/MessageErrorRepositoryTest.java
/*
 * Copyright 2015-2016 USEF Foundation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package energy.usef.core.repository;

import static org.powermock.reflect.Whitebox.setInternalState;

import javax.persistence.EntityManager;
import javax.persistence.EntityManagerFactory;
import javax.persistence.Persistence;
import javax.persistence.PersistenceException;

import org.joda.time.LocalDate;
import org.junit.After;
import org.junit.AfterClass;
import org.junit.Assert;
import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.powermock.modules.junit4.PowerMockRunner;

import energy.usef.core.model.Message;
import energy.usef.core.model.MessageError;

/**
 * JUnit test for the OutgoingMessageErrorRepository class.
 */
@RunWith(PowerMockRunner.class)
public class MessageErrorRepositoryTest {

    /** The factory that produces entity manager. */
    private static EntityManagerFactory entityManagerFactory;
    /** The entity manager that persists and queries the DB. */
    private static EntityManager entityManager;

    private MessageErrorRepository repository;

    @BeforeClass
    public static void initTestFixture() throws Exception {
        // Get the entity manager for the tests.
        entityManagerFactory = Persistence.createEntityManagerFactory("test");
        entityManager = entityManagerFactory.createEntityManager();
    }

    /**
     * Cleans up the session.
     */
    @AfterClass
    public static void closeTestFixture() {
        entityManager.close();
        entityManagerFactory.close();
    }

    @Before
    public void before() {
        repository = new MessageErrorRepository();
        repository.setEntityManager(entityManager);
        setInternalState(repository, "entityManager", entityManager);
        entityManager.getTransaction().begin();
    }

    @After
    public void after() {
        if (entityManager.getTransaction().isActive()) {
            entityManager.getTransaction().rollback();
        }
    }

    /**
     * Tests MessageService.storeMssagError method with null Message.
     */
    @Test(expected = PersistenceException.class)
    public void createOutgoingMessageErrorWithNullMessageTest() {
        MessageError messageError = createMessageError(null);
        repository.persist(messageError);
    }

    /*
     * Fills the required properties
     */
    public static MessageError createMessageError(Message message) {
        MessageError messageError = new MessageError();
        messageError.setMessage(message);
        return messageError;
    }

    @Test
    public void testCleanup() {
        Assert.assertEquals("Expected no deleted objects", 0, repository.cleanup(new LocalDate()));
        Assert.assertEquals("Expected deleted objects", 1, repository.cleanup(new LocalDate("2014-11-22")));
        Assert.assertEquals("Expected no deleted objects", 0, repository.cleanup(new LocalDate("2014-11-22")));
    }


}
